requirements.txt
Django==1.11.3
django-filter==1.0.4
djangorestframework==3.6.3
Markdown==2.6.8
pytz==2017.2
studentit-bookit-client>=0.1,<1.0
